removes old integration tests
